requirements.txt
Django==1.11.6
django-guardian==1.4.9
django-jsonfield==1.0.1
django-notify-x==0.1.8
django-recaptcha==1.3.1
django-taggit==0.22.1
django-userena==2.0.1
easy-thumbnails==2.4.2
html2text==2014.12.29
olefile==0.44
Pillow==4.3.0
pytz==2017.2
six==1.11.0
django-widget-tweaks==1.4.1
